Refactor the plugin registry out of the application
